* make it compatible with x64 app version

+++ libblink/render/geometry/BoxGeometry.cpp
@@ -58,7 +58,7 @@ namespace blink
         {
             std::vector<VertexPos3Uv2Normal> verts;
             buildBox(verts, indis, width, height, depth, widthSegments, heightSegments, depthSegments);
-            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), sizeof(verts[0])*verts.size(), GL_STATIC_DRAW);
+            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), static_cast<uint32>(sizeof(verts[0])*verts.size()), GL_STATIC_DRAW);
         }
         break;
         case BufferAttributes::StockAttributes::Pos3Uv2NormalTangent:
@@ -66,7 +66,7 @@ namespace blink
             std::vector<VertexPos3Uv2NormalTangent> verts;
             buildBox(verts, indis, width, height, depth, widthSegments, heightSegments, depthSegments);
             GeometryUtils::buildTangent(verts, indis);
-            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), sizeof(verts[0])*verts.size(), GL_STATIC_DRAW);
+            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), static_cast<uint32>(sizeof(verts[0])*verts.size()), GL_STATIC_DRAW);
         }
         break;
         default:
@@ -74,7 +74,7 @@ namespace blink
             break;
         }
 
-        uploadIndexBuffer(indis.data(), indis.size(), GL_STATIC_DRAW);
+        uploadIndexBuffer(indis.data(), static_cast<uint32>(indis.size()), GL_STATIC_DRAW);
     }
 
     BoxGeometry::~BoxGeometry()
