+++ libblink/render/geometry/PlaneGeometry.cpp
@@ -49,7 +49,7 @@ namespace blink
         // setup index buffer
         std::vector<uint16> indis;
         GeometryUtils::buildPlaneFaceIndex(indis, 0, widthSegments, heightSegments);
-        uploadIndexBuffer(indis.data(), indis.size(), GL_STATIC_DRAW);
+        uploadIndexBuffer(indis.data(), static_cast<uint32>(indis.size()), GL_STATIC_DRAW);
 
         // setup vertex buffer
         switch (vertexAttribute)
@@ -58,7 +58,7 @@ namespace blink
         {
             std::vector<VertexPos3Uv2Normal> verts;
             buildPlane(verts, width, height, facing, widthSegments, heightSegments);
-            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), sizeof(verts[0])*verts.size(), GL_STATIC_DRAW);
+            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), static_cast<uint32>(sizeof(verts[0])*verts.size()), GL_STATIC_DRAW);
         }
         break;
         case BufferAttributes::StockAttributes::Pos3Uv2NormalTangent:
@@ -66,7 +66,7 @@ namespace blink
             std::vector<VertexPos3Uv2NormalTangent> verts;
             buildPlane(verts, width, height, facing, widthSegments, heightSegments);
             GeometryUtils::buildTangent(verts, indis);
-            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), sizeof(verts[0])*verts.size(), GL_STATIC_DRAW);
+            uploadVertexBuffer(BufferAttributes::fromStock(vertexAttribute), verts.data(), static_cast<uint32>(sizeof(verts[0])*verts.size()), GL_STATIC_DRAW);
         }
         break;
         default:
